requirements.txt
# Python dependencies
anthropic>=0.39.0
python-dotenv>=1.0.0
openai-whisper>=20231117
pyttsx3>=2.90
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
pydub>=0.25.1
# webrtcvad>=2.0.10  # Optional - requires C++ build tools on Windows
